tests/wpt/web-platform-tests/app-history/navigate/navigate-rejection-order-invalidurl-unserializablestate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i"></iframe>
<script>
promise_test(async t => {
  let iframe_constructor = i.contentWindow.DOMException;
  let i_win = i.contentWindow;

  i_win.appHistory.onnavigate = t.unreached_func("onnavigate");
  await promise_rejects_dom(t, "SyntaxError", iframe_constructor, i_win.appHistory.navigate("https://example.com\u0000mozilla.org", { state: document.body }));
}, `navigate() with an invalid URL and unserializable state throws "SyntaxError", not "DataCloneError"`);
</script>
